% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_langages.R
\name{available.languages}
\alias{available.languages}
\title{Return the list of available languages for iatgen translation}
\usage{
available.languages()
}
\value{
Returns dataframe listing the supported languages.

\describe{
\item{Code}{Translation identifier}
\item{Description}{Verbose translation description}
}
}
\description{
Return the list of available languages for iatgen translation
}
\examples{
available.languages()

}
