% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_template.R
\name{export.template}
\alias{export.template}
\title{Export blank translation template}
\usage{
export.template(src_lang = "en")
}
\arguments{
\item{src_lang}{Source language -- "en" for English is the only supported one.}
}
\value{
The path to the template file (class: character). An error is raised
if no template is shipped for \code{src_lang}.
}
\description{
Export blank translation template
}
\examples{
export.template()

}
