% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_template.R
\name{export_template}
\alias{export_template}
\title{Export blank translation template}
\usage{
export_template(src_lang = "en")
}
\arguments{
\item{src_lang}{Source language -- "en" for english is the only supported one.}
}
\description{
Export blank translation template
}
