% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_langs.R
\docType{data}
\name{langs}
\alias{langs}
\title{List of languages included in tr.iatgen}
\format{
A data frame with 2 column(s)
\describe{
\item{Code}{Translation identifier}
\item{Description}{Verbose translation description}
}
}
\usage{
langs
}
\description{
The data.frame containing the List of languages included in tr.iatgen.
}
\keyword{datasets}
