% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runApp}
\alias{runApp}
\title{Run a standalone Shiny App for iatgen translation}
\usage{
runApp(...)
}
\arguments{
\item{..}{All parameters are passed to Shiny runApp function}
}
\description{
Run a standalone Shiny App for iatgen translation
}
