% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_qsf.R
\name{translate_qsf}
\alias{translate_qsf}
\title{Title}
\usage{
translate_qsf(lang, src_lang = "en")
}
\arguments{
\item{lang}{Target language (abbreviation).}

\item{src_lang}{Source language -- "en" for english is the only supported one.}
}
\description{
Title
}
