% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_lang.R
\name{validate_lang_src}
\alias{validate_lang_src}
\title{Title}
\usage{
validate_lang_src(file, src_lang = "en")
}
\arguments{
\item{file}{Source CSV file to validate}

\item{src_lang}{Source language -- "en" for english is the only supported one.}
}
\description{
Title
}
