% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_langages.R
\name{available.languages}
\alias{available.languages}
\title{Return the list of available languages for iatgen translation}
\usage{
available.languages()
}
\description{
Return the list of available languages for iatgen translation
}
