% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langs.R
\docType{data}
\name{langs}
\alias{langs}
\title{Dataset list of included languages in tr.iatgen.}
\format{
A data frame with 2 column(s)
\describe{
\item{langid}{Translation identifier}
\item{description}{Verbose translation description}
...
}
}
\usage{
langs
}
\description{
List of included languages in tr.iatgen.
}
\keyword{datasets}
