% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_qsf_gui.R
\name{translate.qsf.gui}
\alias{translate.qsf.gui}
\title{Graphical user interface (GUI) to translate iatgen generated QSF file}
\usage{
translate.qsf.gui()
}
\value{
Translated file path/location (class: character)
}
\description{
This function allows you to translate an iatgen QSF file by visually
selecting the files and settings for the translation.

It's a wrapper around 'translate.qsf()', meant to be make it more
user-friendly for less technical users.

See the official iatgen website for a web based alternative.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
translate.qsf.gui()
\dontshow{\}) # examplesIf}
}
