% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_qsf.R
\name{translate.qsf}
\alias{translate.qsf}
\title{Title}
\usage{
translate.qsf(file, lang, lang_file = NULL, src_lang = "en", dst_file = NULL)
}
\arguments{
\item{file}{qsf file}

\item{lang}{Target language (abbreviation).}

\item{src_lang}{Source language -- "en" for english is the only supported one.}

\item{dst_file}{save the translated file as. If NULL temporary file will be created.}
}
\value{
translated file location
}
\description{
Title
}
