% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_language.R
\name{validate.language}
\alias{validate.language}
\title{Title}
\usage{
validate.language(file, src_lang = "en")
}
\arguments{
\item{file}{Source CSV file to validate}

\item{src_lang}{Source language -- "en" for english is the only supported one.}
}
\value{
heading of the second column (langugate identifier) if successful, NULL if error
}
\description{
Title
}
